requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.28.75
botocore==1.31.75
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==15.1.0
coverage==7.2.7
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==3.2.22
django-allauth==0.58.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.7.0
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyscopg2==66.0.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.17.0
zope.interface==6.1
